# specs/3-vla-robotics/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Physical AI & Humanoid Robotics RAG Chatbot API
  description: API for the RAG (Retrieval-Augmented Generation) chatbot system integrated with the Physical AI & Humanoid Robotics textbook
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.yourdomain.com
    description: Production server
paths:
  /api/v1/embed:
    post:
      summary: Create embeddings for new content
      description: Process and store new content for RAG retrieval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbedRequest'
      responses:
        '200':
          description: Embeddings created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbedResponse'
        '400':
          description: Invalid request parameters
        '500':
          description: Internal server error
...
  /api/v1/query:
    post:
      summary: Query the RAG system
      description: Retrieve relevant content and generate response using LLM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Query processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Invalid request parameters
        '500':
          description: Internal server error
...
components:
  schemas:
    EmbedRequest:
      type: object
      required:
        - content
        - documentId
        - metadata
      properties:
        content:
          type: string
          description: The content to create embeddings for
        documentId:
          type: string
          description: Unique identifier for the document
        metadata:
          type: object
          description: Additional metadata for the document
          additionalProperties: true
...
    EmbedResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the embedding was successful
        documentId:
          type: string
          description: The document ID for which embeddings were created
        chunksProcessed:
          type: integer
          description: Number of text chunks processed
        createdAt:
          type: string
          format: date-time
          description: When the embeddings were created
    QueryRequest:
      type: object
      required:
        - question
        - sessionId
      properties:
        question:
          type: string
          description: The question to ask
        sessionId:
          type: string
          description: The chat session ID
        maxResults:
          type: integer
          default: 5
          description: Maximum number of results to retrieve from vector store
        contextLength:
          type: integer
          default: 4000
          description: Maximum context length for the LLM
...
    QueryResponse:
      type: object
      required:
        - response
        - sources
        - sessionId
      properties:
        response:
          type: string
          description: The AI-generated response
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
          description: Sources used to generate the response
        sessionId:
          type: string
          description: The chat session ID
        timestamp:
          type: string
          format: date-time
          description: When the response was generated
    Source:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the source
        title:
          type: string
          description: Title of the source
        content:
          type: string
          description: Excerpt from the source
        score:
          type: number
          format: float
          description: Relevance score (0-1)
        url:
          type: string
          description: URL to the source if applicable
...